Bartlet first release.
We have fixed a lot of things up since the initial functionality commit, and now
have a relatively stable and useful codebase.

There is a new concept of suites, which are each loaded automatically when
Bartlet starts by sourcing the Suite's Suite.sh file (for example
StandardSuite/StandardSuite.sh). In the case of StandardSuite, this goes a step
further and sources all the other plugins in the StandardSuite directory.

The README file has all of the documentation and other information that
currnetly exists for Bartlet and is formatted as markdown (expect a rename
momentarily so github will recognize it appropriately).