Refactored the color code.
Refactored the `bartlet_color` function to be closer to git's color
functionality, however we allow you to specify foreground and background
colors as well as one or mode modifiers. All parameters are optional,
however `bartlet_color` only does useful things if at least 1 parameter is
given.
Added `bartlet_color_wrap` which calls directly in to `bartlet_color` with
all arguments and then wraps the output in \[ and \] to support bash's
prompt line requirements.